✨ RENDERER: Canvas Image Preload
💡 What: Implemented asset preloading (images, fonts, backgrounds) for CanvasStrategy using a shared utility extracted from DomStrategy.
🎯 Why: To prevent blank frames or artifacts in Canvas-based compositions that rely on DOM assets.
📊 Impact: Ensures robust rendering for mixed Canvas/DOM workflows.
🔬 Verification: Verified with `tests/verify-canvas-preload.ts` and `tests/verify-enhanced-dom-preload.ts`.
