+++ docs/PROGRESS-RENDERER.md
@@ -1,5 +1,8 @@
 # Renderer Progress Log
 
+## RENDERER v1.76.0
+- ✅ Completed: Canvas Image Preload - Implemented comprehensive asset preloading for `CanvasStrategy` (images, fonts, backgrounds) using a shared utility extracted from `DomStrategy`. Verified with `verify-canvas-preload.ts`.
+
 ## RENDERER v1.75.2
 - ✅ Completed: Restore Development Environment - Restored dependencies and installed Playwright browsers, enabling successful execution of the verification suite and resolving the environment blockage.
